fix: 未ログイン時のトップページをサーバーレンダリングしLPを初期HTMLに含める
これまで / の初期HTMLは認証確認中のスピナーのみで、クローラーには
ほぼ空のページに見えていた（AdSense「有用性の低いコンテンツ」の主要因）。

- LP部分を GuestLanding サーバーコンポーネントとして切り出し
- / はSupabaseの認証Cookieの有無でサーバー側分岐:
  Cookieなし → GuestLanding をSSR（初期HTMLにLP全文）
  Cookieあり → 従来のクライアントホーム（挙動変更なし）
- Cookieはあるがセッション無効の場合は従来どおりクライアント側で
  GuestLanding にフォールバック
